+++ llm_mcp_server/tests/__init__.py
@@ -1,2 +1,3 @@
 from . import test_mcp_handshake
 from . import test_mcp_oauth
+from . import test_mcp_tool_exposure